Logback Implementation

+++ employeeInterface/src/br/com/employeInterface/employeeInterface/EmployeeInterface.java
@@ -4,7 +4,7 @@ public interface EmployeeInterface {
     void createEmployee (String employeeFirstName, String employeeMiddleName, String employeeLastname, String employeeIdNumber, String employeeCpfNumber, int chooseJobe_title, double employeeSalary, String hireDate);
     void terminationDateEmployee(long employeeEnrollmentNumber, String terminationDateEmployee);
     void findEmployee(int employeeEnrollmentNumber);
-    void deleteEmployee(int employeeEnrollmentNumber);
+    void deleteEmployee(long employeeEnrollmentNumber);
     void  calculateTenure(long employeeEnrollmentNumber);
     void showJobe_title();
     void printEmployeeList();
